# docs/swagger/openapi.yaml
...
openapi: 3.0.3
info:
  title: Foundations
  version: '1.0'
  description: The API for the Foundations project
servers:
  - url: http://localhost:8081
  - url: http://localhost:8080
paths:
...
  "/api/v1/foundations/{id}":
    put:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Foundation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/FoundationPost"
          text/json:
            schema:
              "$ref": "#/components/schemas/FoundationPost"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/FoundationPost"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Foundation"
            application/json:
              schema:
                "$ref": "#/components/schemas/Foundation"
            text/json:
              schema:
                "$ref": "#/components/schemas/Foundation"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Foundation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Foundation"
            application/json:
              schema:
                "$ref": "#/components/schemas/Foundation"
            text/json:
              schema:
                "$ref": "#/components/schemas/Foundation"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Foundation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Foundation"
            application/json:
              schema:
                "$ref": "#/components/schemas/Foundation"
            text/json:
              schema:
                "$ref": "#/components/schemas/Foundation"
        '404':
          description: Not Found
  "/api/v1/foundations/{id}/donate":
    post:
        security:
          [{
            bearerAuth: []
        }]
        tags:
        - Foundation
        parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        requestBody:
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FoundationDonate"
            text/json:
              schema:
                "$ref": "#/components/schemas/FoundationDonate"
            application/*+json:
              schema:
                "$ref": "#/components/schemas/FoundationDonate"
        responses:
          '200':
            description: Success
            content:
              text/plain:
                schema:
                  "$ref": "#/components/schemas/Transaction"
              application/json:
                schema:
                  "$ref": "#/components/schemas/Transaction"
              text/json:
                schema:
                  "$ref": "#/components/schemas/Transaction"
          '400':
            description: Bad Request
          '401':
            description: Unauthorized
          '404':
            description: Not Found
          '409':
            description: Conflict
...
  "/api/v1/foundrisings/{id}":
    put:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Foundrising
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/FoundrisingPut"
          text/json:
            schema:
              "$ref": "#/components/schemas/FoundrisingPut"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/FoundrisingPut"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            application/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            text/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    patch:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Foundrising
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/FoundrisingPut"
          text/json:
            schema:
              "$ref": "#/components/schemas/FoundrisingPut"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/FoundrisingPut"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            application/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            text/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Foundrising
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            application/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            text/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Foundrising
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            application/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
            text/json:
              schema:
                "$ref": "#/components/schemas/Foundrising"
        '404':
          description: Not Found
...
  "/api/v1/transactions/{id}":
    get:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Transaction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Transaction"
            application/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
            text/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    delete:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Transaction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Transaction"
            application/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
            text/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
...
components:
  schemas:
    Foundation:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        curFoudrisingAmount:
          type: integer
          format: int32
        closedFoundrisingAmount:
          type: integer
          format: int32
        volunteer_amount:
          type: integer
          format: int32
        fund_balance:
          type: number
          format: double
        income:
          type: number
          format: double
        outcome:
          type: number
          format: double
        country:
          type: string
          nullable: true
        login:
          type: string
          nullable: true
      additionalProperties: false
    FoundationPost:
      type: object
      properties:
        name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        login:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    FoundationDonate:
      type: object
      properties:
        foundrising_id:
          type: string
          nullable: true
        sum_of_money:
          type: integer
          format: int32
        comm:
          type: string
          nullable: true
      additionalProperties: false
    Foundrising:
      type: object
      properties:
        id:
          type: integer
          format: int32
        found_id:
          type: integer
          format: int32
        reqSum:
          type: integer
          format: int32
        ph_amount:
          type: integer
          format: int32
        curSum:
          type: integer
          format: int32
        description:
          type: string
          nullable: false
        creation_date:
          type: string
          nullable: false
        close_date:
          type: string
          nullable: false
      additionalProperties: false
...
    FoundrisingPut:
      type: object
      properties:
        reqSum:
          type: integer
          format: int32
        description:
          type: string
          nullable: false
      additionalProperties: false
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        fromId:
          type: integer
          format: int32
          nullable: true
        toId:
          type: integer
          format: int32
          nullable: true
        fromType:
          type: string
          nullable: true
        toType:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        sum:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT